+++ spica-server/src/main/java/org/spica/server/software/api/SoftwareController.java
@@ -3,7 +3,7 @@ package org.spica.server.software.api;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
-import org.spica.server.software.model.Software;
+import org.spica.server.software.model.SoftwareInfo;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,8 +12,8 @@ import org.springframework.web.bind.annotation.RestController;
 @CrossOrigin(origins="*")
 public class SoftwareController implements SoftwareApi {
 
-  public ResponseEntity<List<Software>> getSoftware() {
-    Software software = new Software();
+  public ResponseEntity<List<SoftwareInfo>> getSoftware() {
+    SoftwareInfo software = new SoftwareInfo();
     software.setName("Software");
     software.setDescription("Description");
     software.setId("id");