+++ spica-server/src/main/java/org/spica/server/software/db/SoftwareRepository.java
@@ -1,10 +1,11 @@
 package org.spica.server.software.db;
 
+import org.spica.server.software.domain.Software;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-interface SoftwareRepository extends CrudRepository<SoftwareDb, Long> {
+interface SoftwareRepository extends CrudRepository<Software, Long> {
 
 
 }