+++ spica-server/src/main/java/org/spica/server/software/domain/SoftwarePersistencePort.java
@@ -1,9 +1,0 @@
-package org.spica.server.software.domain;
-
-import org.spica.server.software.model.SoftwareAggregate;
-
-public interface SoftwarePersistencePort {
-
-  SoftwareAggregate findAll ();
-
-}