+++ spica-server/src/test/java/org/spica/server/software/db/SoftwarePersistenceAdapterTest.java
@@ -1,4 +1,0 @@
-package org.spica.server.software.db;
-
-public class SoftwarePersistenceAdapterTest {
-}